+++ server/src/knowledge-bases/dto/knowledge-base.dto.ts
@@ -25,10 +25,6 @@ export class CreateKnowledgeBaseDto {
 
   @IsString()
   @IsOptional()
-  code?: string;
-
-  @IsString()
-  @IsOptional()
   description?: string;
 
   @IsString()
@@ -150,14 +146,6 @@ export class QueryNextCategoryCodeDto {
   @Min(1)
   @IsOptional()
   parentId?: number;
-}
-
-export class QueryNextBaseCodeDto {
-  @Type(() => Number)
-  @IsInt()
-  @Min(1)
-  @IsOptional()
-  categoryId?: number;
 }
 
 export class CreateKnowledgeBaseDocumentDto {